+++ CHANGES.md
@@ -1,12 +1,11 @@
 **Added**  
-* untested method for getting an if statement from an SQFStatement
+* 
 
 **Changed**  
 * 
 
 **Fixed**  
-* 
+* renaming SQFVariable 
 
 **Notes**
-* fix renaming variable
 * scope works, but it doesn't work for deep nested scopes